loading a file into an 3d geometry class and rendering it with texture

+++ OpenGLFromScratch/OpenGLFromScratch/src/Geometry/G3DSimpleGeometry3D.h
@@ -25,7 +25,7 @@ public:
 	std::vector<GLfloat> Vertices;
 	std::vector<GLfloat> Normals;
 	std::vector<GLfloat> UV;
-	std::vector<GLuint> Indices;
+	std::vector<GLushort> Indices;
 
 	G3DMesh3D();
 };
